update docker entry file to handle server settings
This added quite a bit of bulk into the docker entry file since the
settings directory isn't created at runtime but instead exists by
default. Therefore the server won't just create the settings directory
if we symlink it into the server config directory. Instead we have to
handle the moving, and symlinking of it ourselves.